refactor: merge user interfaces to a single file user.go, permission interface files to a single file permission.go, and role interface files to a single file role.go
