--- internal/core/ports/permission_handler.go
@@ -1,10 +1,0 @@
-package ports
-
-import "github.com/gin-gonic/gin"
-
-type PermissionHandler interface {
-	CreatePermission(c *gin.Context)
-	DeletePermission(c *gin.Context)
-	GetAllPermissions(c *gin.Context)
-	GetPermissionByID(c *gin.Context)
-}
